+++ app/src/main/java/net/pilseong/todocompose/data/model/Photo.kt
@@ -9,10 +9,10 @@ import java.time.ZonedDateTime
 @Entity(tableName = PHOTO_TABLE)
 data class Photo  constructor(
     @PrimaryKey(autoGenerate = true)
-    val id: Int = 0,
+    val id: Long = 0,
     val uri: String,
     val filename: String,
-    val memoId: Int,
+    val memoId: Long,
     @ColumnInfo(name = "created_at")
     val createdAt: ZonedDateTime = ZonedDateTime.now(),
     @ColumnInfo(name = "updated_at")
